Moved TestTradable to its own package.
