+++ src/test/scala/org/economicsl/auctions/TestTradable.scala
@@ -13,9 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package org.economicsl.auctions.singleunit
-
-import org.economicsl.auctions.{Currency, Tradable}
+package org.economicsl.auctions
 
 
 case class TestTradable(tick: Currency) extends Tradable
